fix(Core): indexation of current graphic lib was initiate to -1 because he couldn't found it Now if it is -1 we thro ERROR: not in lib dir + Add many features in ncurses by showing everything with is current color

+++ src/core/ArcadeCore.hpp
@@ -32,8 +32,8 @@ namespace arcade {
             std::string currentLib;
             std::unique_ptr<IGraphic> graphic_;
             std::unique_ptr<IGame> game_;
-            size_t gameIndex;
-            size_t graphicIndex;
+            int gameIndex;
+            int graphicIndex;
     };
 }
